requirements.txt
django
pycryptodome
googlemaps

